fix: hero card trigger from profile nav bar
