+++ src/modules/profile/state/heroCardAtom.ts
@@ -1,0 +1,11 @@
+import { atom } from 'jotai'
+
+import { UserForProfilePageFragment, UserHeroStats } from '@/types'
+
+type HeroCardAtomType = {
+  isOpen: boolean
+  user: UserForProfilePageFragment
+  stats: UserHeroStats
+}
+
+export const heroCardAtom = atom<HeroCardAtomType | null>(null)
